+++ src/main/java/org/example/n1Exe1/Main.java
@@ -1,4 +1,4 @@
-package n1Exe1;
+package org.example.n1Exe1;
 public class Main {
 
     public static void main (String args []){
